backend/requirements.txt
firebase_admin==6.1.0
Flask==2.3.1
nltk==3.8.1
numpy==1.23.5
pandas==2.0.1
pdfplumber==0.9.0
Pillow==9.5.0
Pillow==9.5.0
PyPDF2==3.0.1
Pyrebase==3.0.27
pytesseract==0.3.10
Requests==2.29.0
scikit_learn==1.2.2
tensorflow==2.12.0
urllib3==1.26.15
Wand==0.6.11
